Renamed and fixed minor bugs
